+++ apps/desktop/src/lib/prosemark-core/markdown/index.ts
@@ -4,6 +4,7 @@ import { additionalMarkdownSyntaxTags } from "../syntaxHighlighting";
 import { frontmatterMarkdownSyntaxExtension } from "./frontmatter";
 import { nestedLinkAsPlainText } from "./nestedLinkAsPlainText";
 import { mathMarkdownSyntaxExtension } from "./mathMarkdown";
+import { spaceDestinationLinksMarkdownSyntaxExtension } from "./spaceDestinationLinks";
 
 export { markdownTags } from "./tags";
 export {
@@ -16,10 +17,12 @@ export { escapeMarkdownSyntaxExtension } from "../hide";
 export { additionalMarkdownSyntaxTags } from "../syntaxHighlighting";
 export { emojiMarkdownSyntaxExtension, dashMarkdownSyntaxExtension } from "../fold";
 export { mathDelimiterTag, mathFormulaTag, mathMarkdownSyntaxExtension } from "./mathMarkdown";
+export { spaceDestinationLinksMarkdownSyntaxExtension } from "./spaceDestinationLinks";
 
 export const prosemarkMarkdownSyntaxExtensions = [
   additionalMarkdownSyntaxTags,
   frontmatterMarkdownSyntaxExtension,
+  spaceDestinationLinksMarkdownSyntaxExtension,
   nestedLinkAsPlainText,
   escapeMarkdownSyntaxExtension,
   emojiMarkdownSyntaxExtension,